fix delete photo button, now works without js, add img montage buttons and styles

+++ inc/edit_photo/delete_photo.php
@@ -8,36 +8,29 @@ include($_SERVER["DOCUMENT_ROOT"]."/config/connect.php");
 /*
 ** DELETE PHOTO FROM SERVER FOLDER
 */
-try {
-	$result = $conn->prepare(
-		"SELECT src FROM user_img WHERE img_id=" .
-		$_POST['img_id']);
-	$result->execute();
-	$file = $result->setFetchMode(PDO::FETCH_ASSOC);
-	$file = $result->fetchAll();
-	$file = str_replace('http://localhost:8080/', '', $file[0]['src']);
-	$file = "../../" . $file;
-	if (!unlink($file))
-		{
-		echo ("Error deleting $file"."\n");
+if (!empty($_GET["img_id"]) || !empty($_GET["pg"])) {
+	try {
+		$result = $conn->prepare(
+			"SELECT src FROM user_img WHERE img_id=" .
+			$_GET['img_id']);
+		$result->execute();
+		$file = $result->setFetchMode(PDO::FETCH_ASSOC);
+		$file = $result->fetchAll();
+		$file = str_replace('http://localhost:8080/', '', $file[0]['src']);
+		$file = "../../" . $file;
+		if (!unlink($file)) {
+			echo ("Error deleting $file"."\n");
 		}
-	else
-		{
-		echo ("Deleted $file"."\n");
-		}
-	// sql to delete a record
-	$sql = "DELETE FROM user_img WHERE img_id=" . 
-			$_POST['img_id'];
+		$sql = "DELETE FROM user_img WHERE img_id=" . 
+				$_GET['img_id'];
 
-	// use exec() because no results are returned
-	$conn->exec($sql);
-	echo "Record deleted successfully";
+		$conn->exec($sql);
+	} catch (PDOException $e) {
+		echo "Unable to retrieved results";
+		exit;
 	}
-catch (PDOException $e) {
-	echo "Unable to retrieved results";
-	exit;
-	}
-
+	header("location:http://localhost:8080/account.php?pg=".$_GET['pg']);
+}
 $conn = null;
 
 ?>
